move dockgraph_cli.py into the dockgraph module
