--- bin/dockgraph
@@ -1,1 +1,0 @@
-dockgraph_cli.py
